Refactoring tiny tools
