+++ plugins/woo-ai/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './shuffleArray';
 export * from './jetpack-completion';
 export * from './recordTracksFactory';
 export * from './get-post-id';
+export * from './tiny-tools';
